refactor(data): テーブル以外でも利用するデータの名前にtablesとつけていたため削除した
