+++ store/pages/data/components/types/aggregationByDow.ts
