+++ store/pages/data/components/types/aggregationByDrinks.ts
